pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "soccer-rig"
version = "1.0.0"
description = "Multi-Camera Pi 5 Soccer Recording System"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Soccer Rig Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Video :: Capture",
]
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "gunicorn>=21.0.0",
    "psutil>=5.9.0",
    "pyyaml>=6.0.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.0.0",
    "netifaces>=0.11.0",
    "zeroconf>=0.131.0",
    "requests>=2.31.0",
    "python-dateutil>=2.8.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
audio = ["simpleaudio>=1.0.4", "numpy>=1.24.0"]
dev = ["pytest>=7.0.0", "pytest-cov>=4.0.0", "black>=23.0.0", "flake8>=6.0.0"]

[project.scripts]
soccer-rig = "soccer_rig.app:main"

[project.urls]
Homepage = "https://github.com/soccer-rig/soccer-rig"
Documentation = "https://github.com/soccer-rig/soccer-rig#readme"
Issues = "https://github.com/soccer-rig/soccer-rig/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.html", "*.css", "*.js"]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=soccer_rig --cov-report=term-missing"
